Refactor pcm dependencies in builder
