+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/META-INF/MANIFEST.MF
@@ -21,6 +21,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.24.0",
  org.palladiosimulator.dataflow.nodecharacteristics;bundle-version="0.1.0"
 Export-Package: org.palladiosimulator.dataflow.confidentiality.analysis,
  org.palladiosimulator.dataflow.confidentiality.analysis.builder,
+ org.palladiosimulator.dataflow.confidentiality.analysis.builder.pcm,
  org.palladiosimulator.dataflow.confidentiality.analysis.characteristics,
  org.palladiosimulator.dataflow.confidentiality.analysis.entity,
  org.palladiosimulator.dataflow.confidentiality.analysis.entity.pcm,
