+++ bundles/org.palladiosimulator.dataflow.confidentiality.analysis/src/org/palladiosimulator/dataflow/confidentiality/analysis/builder/AnalysisBuilderData.java
@@ -12,13 +12,9 @@ import org.palladiosimulator.dataflow.confidentiality.analysis.resource.PCMURIRe
 import org.palladiosimulator.dataflow.confidentiality.analysis.resource.ResourceLoader;
 
 public class AnalysisBuilderData {
-	private boolean standalone;
-	private boolean legacy;
-	private String modelProjectName;
-	private Class<? extends Plugin> pluginActivator;
-	private String relativeUsageModelPath;
-	private String relativeAllocationModelPath;
-	private String relativeNodeCharacteristicsPath;
+	protected boolean standalone;
+	protected String modelProjectName;
+	
 	
 	/**
 	 * Sets the model porject name
@@ -37,22 +33,6 @@ public class AnalysisBuilderData {
 	}
 	
 	/**
-	 * Sets the plugin activator of the project
-	 * @param pluginActivator Eclipse plugin activator class
-	 */
-	public void setPluginActivator(Class<? extends Plugin> pluginActivator) {
-		this.pluginActivator = pluginActivator;
-	}
-	
-	/**
-	 * Returns the plugin activator of the project
-	 * @return Eclipse plugin activator class of the project
-	 */
-	public Class<? extends Plugin> getPluginActivator() {
-		return pluginActivator;
-	}
-	
-	/**
 	 * Sets the standalone mode of the analysis 
 	 * @param standalone New configured mode of the analysis
 	 */
@@ -67,99 +47,4 @@ public class AnalysisBuilderData {
 	public boolean isStandalone() {
 		return standalone;
 	}
-	
-	/**
-	 * Sets the legacy mode of the analysis to allow the loading of EMF Profiles
-	 * @param legacy New value of the legacy mode
-	 */
-	public void setLegacy(boolean legacy) {
-		this.legacy = legacy;
-	}
-	
-	/**
-	 * Returns, whether or not the analysis is in legacy mode
-	 * @return Returns true, if the analysis is in legacy mode and EMF Profiles are loaded. Otherwise, the method returns false
-	 */
-	public boolean isLegacy() {
-		return legacy;
-	}
-	
-	/**
-	 * Sets the relative path to the usage model used in the analysis
-	 * @param relativeUsageModelPath Relative path to the usage model
-	 */
-	public void setRelativeUsageModelPath(String relativeUsageModelPath) {
-		this.relativeUsageModelPath = relativeUsageModelPath;
-	}
-	
-	/**
-	 * Returns the configured relative path to the usage model of the analysis
-	 * @return Relative path to the usage model
-	 */
-	public String getRelativeUsageModelPath() {
-		return relativeUsageModelPath;
-	}
-	
-	/**
-	 * Sets the relative path to the allocation model used in the analysis
-	 * @param relativeAllocationModelPath Relative path to the allocation model
-	 */
-	public void setRelativeAllocationModelPath(String relativeAllocationModelPath) {
-		this.relativeAllocationModelPath = relativeAllocationModelPath;
-	}
-	
-	/**
-	 * Returns the configured relative path to the allocation model
-	 * @return Relative path to the allocation model
-	 */
-	public String getRelativeAllocationModelPath() {
-		return relativeAllocationModelPath;
-	}
-	
-	/**
-	 * Sets the relative path to the node characteristics model that is used in the analysis
-	 * @param relativeNodeCharacteristicsPath Relative path to the node characteristics model
-	 */
-	public void setRelativeNodeCharacteristicsPath(String relativeNodeCharacteristicsPath) {
-		this.relativeNodeCharacteristicsPath = relativeNodeCharacteristicsPath;
-	}
-	
-	/**
-	 * Returns the relative path to the node characteristics model the analysis is configured to use
-	 * @return Relative path to the node characteristcs model
-	 */
-	public String getRelativeNodeCharacteristicsPath() {
-		return relativeNodeCharacteristicsPath;
-	}
-
-	/**
-	 * Creates a new analysis data object from the configured data. It does not check, whether all parameters are set correctly. Use {@code DataFlowConfidentialityAnalysisBuilder} instead
-	 * @return Returns a new data object for the analysis
-	 */
-	public AnalysisData createAnalysisData() {
-		if (this.isLegacy()) {
-			ResourceLoader resourceLoader = new PCMURIResourceLoader(this.createRelativePluginURI(relativeUsageModelPath, modelProjectName), 
-					this.createRelativePluginURI(relativeAllocationModelPath, modelProjectName), Optional.empty());
-			return new AnalysisData(resourceLoader, 
-					new LegacyPCMNodeCharacteristicsCalculator(resourceLoader), new PCMDataCharacteristicsCalculatorFactory(resourceLoader));
-		} else {
-			ResourceLoader resourceLoader = new PCMURIResourceLoader(this.createRelativePluginURI(relativeUsageModelPath, modelProjectName), 
-					this.createRelativePluginURI(relativeAllocationModelPath, modelProjectName), 
-					Optional.of(this.createRelativePluginURI(relativeNodeCharacteristicsPath, modelProjectName)));
-			return new AnalysisData(resourceLoader, 
-					new PCMNodeCharacteristicsCalculatorImpl(resourceLoader), new PCMDataCharacteristicsCalculatorFactory(resourceLoader));
-		}
-	}
-	
-    
-    /**
-     * Creates a relative plugin uri from the given relative path
-     * @param relativePath Given relative path
-     * @return Returns plugin path with the given project name and provided relative path
-     */
-    private URI createRelativePluginURI(String relativePath, String modelProjectName) {
-        String path = Paths.get(modelProjectName, relativePath)
-            .toString();
-        return URI.createPlatformPluginURI(path, false);
-    }
 }
